Support global configuration of options
