+++ src/main/groovy/com/github/eerohele/SaxonPluginConfiguration.groovy
@@ -1,0 +1,5 @@
+package com.github.eerohele
+
+@Singleton
+class SaxonPluginConfiguration implements OptionsConfig {
+}
