+++ src/main/groovy/com/github/eerohele/SaxonXsltTask.groovy
@@ -22,7 +22,7 @@ import javax.inject.Inject
 import java.nio.file.Path
 
 @SuppressWarnings('MethodCount')
-class SaxonXsltTask extends DefaultTask {
+class SaxonXsltTask extends DefaultTask implements OptionsConfig {
     protected static final String PERIOD = '.'
     protected static final String XSLT_NAMESPACE = 'http://www.w3.org/1999/XSL/Transform'
 
@@ -34,8 +34,6 @@ class SaxonXsltTask extends DefaultTask {
 
     @Classpath
     final ConfigurableFileCollection classpath
-
-    protected final Map<String, String> advancedOptions = [:]
 
     protected Map<String, String> stylesheetParams = [:]
 
@@ -71,66 +69,12 @@ class SaxonXsltTask extends DefaultTask {
 
     // START COMMON OPTIONS
 
-    void config(Object config) {
-        this.options.config = project.file(config)
-    }
-
-    void collectionResolver(String resolver) {
-        this.options.collectionResolver = resolver
-    }
-
-    void catalog(Object catalog) {
-        this.options.catalog = project.file(catalog)
-
-        URI catalogs = this.options.catalog.toURI()
-        catalogManager.setCatalogFiles(catalogs.toString())
-        xmlCatalog = new Catalog(catalogManager)
-        xmlCatalog.setupReaders()
-        xmlCatalog.parseCatalog(catalogs.toString())
-    }
-
-    void dtd(Object dtd) {
-        this.options.dtd = dtd
-    }
-
-    void expand(Object expand) {
-        this.options.expand = expand
-    }
-
-    void explain(Object explain) {
-        this.options.explain = explain
-    }
-
-    void initializer(String initializer) {
-        this.options.initializer = initializer
-    }
-
-    void initialMode(String initialMode) {
-        this.options.initialMode = initialMode
-    }
-
-    void initialTemplate(String initialTemplate) {
-        this.options.initialTemplate = initialTemplate
-    }
-
     void input(Object input) {
         this.options.input = input
     }
 
-    void lineNumbers(Object lineNumbers) {
-        this.options.lineNumbers = lineNumbers
-    }
-
-    void messageReceiver(String receiver) {
-        this.options.messageReceiver = receiver
-    }
-
     void output(Object output) {
         this.options.output = project.file(output)
-    }
-
-    void sourceSaxParser(String parser) {
-        this.options.sourceSaxParser = parser
     }
 
     @SuppressWarnings('CatchException')
@@ -148,22 +92,6 @@ class SaxonXsltTask extends DefaultTask {
         }
     }
 
-    void stylesheetSaxParser(String parser) {
-        this.options.stylesheetSaxParser = parser
-    }
-
-    void suppressJavaCalls(Object suppress) {
-        this.options.suppressJavaCalls = suppress
-    }
-
-    void uriResolver(Object resolver) {
-        this.options.uriResolver = resolver
-    }
-
-    void useAssociatedStylesheet(Object use) {
-        this.options.useAssociatedStylesheet = use
-    }
-
     void outputFileExtension(String extension) {
         this.pluginOptions.outputFileExtension = extension
     }
@@ -171,466 +99,6 @@ class SaxonXsltTask extends DefaultTask {
     void outputDirectoryLayout(String layout) {
         this.pluginOptions.outputDirectoryLayout = layout
     }
-
-    // START ADVANCED OPTIONS
-
-    void allowExternalFunctions(Boolean x) {
-        this.advancedOptions['allow-external-functions'] = x;
-    }
-
-    void allowMultithreading(Boolean x) {
-        this.advancedOptions['allow-multithreading'] = x;
-    }
-
-    void allowOldJavaURIFormat(Boolean x) {
-        this.advancedOptions['allow-old-java-uri-format'] = x;
-    }
-
-    void allowSyntaxExtensions(Boolean x) {
-        this.advancedOptions.allowSyntaxExtensions = x;
-    }
-
-    void assertionsCanSeeComments(Boolean x) {
-        this.advancedOptions.assertionsCanSeeComments = x;
-    }
-
-    void collationUriResolver(Object x) {
-        this.advancedOptions['collation-uri-resolver'] = x;
-    }
-
-    void collationUriResolverClass(Object x) {
-        this.advancedOptions['collation-uri-resolver-class'] = x;
-    }
-
-    void collectionFinder(Object x) {
-        this.advancedOptions['collection-finder'] = x;
-    }
-
-    void collectionFinderClass(Object x) {
-        this.advancedOptions['collection-finder-class'] = x;
-    }
-
-    void collectionUriResolver(Object x) {
-        this.advancedOptions['collection-uri-resolver'] = x;
-    }
-
-    void collectionUriResolverClass(Object x) {
-        this.advancedOptions['collection-uri-resolver-class'] = x;
-    }
-
-    void compileWithTracing(Boolean x) {
-        this.advancedOptions.with = x;
-    }
-
-    void configuration(Object x) {
-        this.advancedOptions.configuration = x;
-    }
-
-    void configurationFile(Object x) {
-        this.advancedOptions['configuration-file'] = x;
-    }
-
-    void debugByteCode(Object x) {
-        this.advancedOptions.debugByteCode = x;
-    }
-
-    void debugByteCodeDir(Object x) {
-        this.advancedOptions.debugByteCodeDir = x;
-    }
-
-    void defaultCollation(Object x) {
-        this.advancedOptions.defaultCollation = x;
-    }
-
-    void defaultCollection(Object x) {
-        this.advancedOptions.defaultCollection = x;
-    }
-
-    void defaultCountry(Object x) {
-        this.advancedOptions.defaultCountry = x;
-    }
-
-    void defaultLanguage(Object x) {
-        this.advancedOptions.defaultLanguage = x;
-    }
-
-    void defaultRegexEngine(Object x) {
-        this.advancedOptions.defaultRegexEngine = x;
-    }
-
-    void disableXslEvaluate(Boolean x) {
-        this.advancedOptions.disableXslEvaluate = x;
-    }
-
-    void displayByteCode(Boolean x) {
-        this.advancedOptions.displayByteCode = x;
-    }
-
-    void validation(Object x) {
-        this.advancedOptions.validation = x;
-    }
-
-    void dtdValidationRecoverable(Object x) {
-        this.advancedOptions['dtd-validation-recoverable'] = x;
-    }
-
-    void eagerEvaluation(Object x) {
-        this.advancedOptions.eagerEvaluation = x;
-    }
-
-    void entityResolverClass(Object x) {
-        this.advancedOptions.entityResolverClass = x;
-    }
-
-    void environmentVariableResolver(Object x) {
-        this.advancedOptions.environmentVariableResolver = x;
-    }
-
-    void environmentVariableResolverClass(Object x) {
-        this.advancedOptions.environmentVariableResolverClass = x;
-    }
-
-    void errorListenerClass(Object x) {
-        this.advancedOptions.errorListenerClass = x;
-    }
-
-    void expathFileDeleteTemporaryFiles(Object x) {
-        this.advancedOptions.expathFileDeleteTemporaryFiles = x;
-    }
-
-    void generateByteCode(Object x) {
-        this.advancedOptions.generateByteCode = x;
-    }
-
-    void ignoreSAXSourceParser(Object x) {
-        this.advancedOptions.ignoreSAXSourceParser = x;
-    }
-
-    void implicitSchemaImports(Object x) {
-        this.advancedOptions.implicitSchemaImports = x;
-    }
-
-    void lazyConstructionMode(Object x) {
-        this.advancedOptions.lazyConstructionMode = x;
-    }
-
-    void licenseFileLocation(Object x) {
-        this.advancedOptions.licenseFileLocation = x;
-    }
-
-    void markDefaultedAttributes(Object x) {
-        this.advancedOptions.markDefaultedAttributes = x;
-    }
-
-    void maxCompiledClasses(Object x) {
-        this.advancedOptions.maxCompiledClasses = x;
-    }
-
-    void messageEmitterClass(Object x) {
-        this.advancedOptions.messageEmitterClass = x;
-    }
-
-    void moduleURIResolver(Object x) {
-        this.advancedOptions.moduleURIResolver = x;
-    }
-
-    void moduleURIResolverClass(Object x) {
-        this.advancedOptions.moduleURIResolverClass = x;
-    }
-
-    void monitorHotSpotByteCode(Object x) {
-        this.advancedOptions.monitorHotSpotByteCode = x;
-    }
-
-    void multipleSchemaImports(Object x) {
-        this.advancedOptions.multipleSchemaImports = x;
-    }
-
-    void namePool(Object x) {
-        this.advancedOptions.namePool = x;
-    }
-
-    void occurrenceLimits(Object x) {
-        this.advancedOptions.occurrenceLimits = x;
-    }
-
-    void optimizationLevel(Object x) {
-        this.advancedOptions.optimizationLevel = x;
-    }
-
-    void outputURIResolver(Object x) {
-        this.advancedOptions.outputURIResolver = x;
-    }
-
-    void outputURIResolverClass(Object x) {
-        this.advancedOptions.outputURIResolverClass = x;
-    }
-
-    void parserFeature(Object x) {
-        this.advancedOptions.parserFeature = x;
-    }
-
-    void parserProperty(Object x) {
-        this.advancedOptions.parserProperty = x;
-    }
-
-    void preEvaluateDocFunction(Object x) {
-        this.advancedOptions.preEvaluateDocFunction = x;
-    }
-
-    void preferJaxpParser(Object x) {
-        this.advancedOptions.preferJaxpParser = x;
-    }
-
-    void recognizeUriQueryParameters(Object x) {
-        this.advancedOptions['recognize-uri-query-parameters'] = x;
-    }
-
-    void recoveryPolicy(Object x) {
-        this.advancedOptions.recoveryPolicy = x;
-    }
-
-    void recoveryPolicyName(Object x) {
-        this.advancedOptions.recoveryPolicyName = x;
-    }
-
-    void resultDocumentThreads(Object x) {
-        this.advancedOptions.resultDocumentThreads = x;
-    }
-
-    void retainDtdAttributeTypes(Object x) {
-        this.advancedOptions['retain-dtd-attribute-types'] = x;
-    }
-
-    void schemaURIResolver(Object x) {
-        this.advancedOptions.schemaURIResolver = x;
-    }
-
-    void schemaURIResolverClass(Object x) {
-        this.advancedOptions.schemaURIResolverClass = x;
-    }
-
-    void schemaValidation(Object x) {
-        this.advancedOptions['schema-validation'] = x;
-    }
-
-    void schemaValidationMode(Object x) {
-        this.advancedOptions['schema-validation-mode'] = x;
-    }
-
-    void serializerFactoryClass(Object x) {
-        this.advancedOptions.serializerFactoryClass = x;
-    }
-
-    void sourceParserClass(Object x) {
-        this.advancedOptions.sourceParserClass = x;
-    }
-
-    void sourceResolverClass(Object x) {
-        this.advancedOptions.sourceResolverClass = x;
-    }
-
-    void stableCollectionUri(Object x) {
-        this.advancedOptions.stableCollectionUri = x;
-    }
-
-    void stableUnparsedText(Object x) {
-        this.advancedOptions.stableUnparsedText = x;
-    }
-
-    void standardErrorOutputFile(Object x) {
-        this.advancedOptions.standardErrorOutputFile = x;
-    }
-
-    void streamability(Object x) {
-        this.advancedOptions.streamability = x;
-    }
-
-    void strictStreamability(Object x) {
-        this.advancedOptions.strictStreamability = x;
-    }
-
-    void streamingFallback(Object x) {
-        this.advancedOptions.streamingFallback = x;
-    }
-
-    void stripWhitespace(Object x) {
-        this.advancedOptions['strip-whitespace'] = x;
-    }
-
-    void styleParserClass(Object x) {
-        this.advancedOptions.styleParserClass = x;
-    }
-
-    void suppressEvaluationExpiryWarning(Object x) {
-        this.advancedOptions.suppressEvaluationExpiryWarning = x;
-    }
-
-    void suppressXPathWarnings(Object x) {
-        this.advancedOptions.suppressXPathWarnings = x;
-    }
-
-    void suppressXsltNamespaceCheck(Object x) {
-        this.advancedOptions.suppressXsltNamespaceCheck = x;
-    }
-
-    void thresholdForCompilingTypes(Object x) {
-        this.advancedOptions.thresholdForCompilingTypes = x;
-    }
-
-    void timing(Object x) {
-        this.advancedOptions.timing = x;
-    }
-
-    void traceExternalFunctions(Object x) {
-        this.advancedOptions['trace-external-functions'] = x;
-    }
-
-    void traceListener(Object x) {
-        this.advancedOptions.traceListener = x;
-    }
-
-    void traceListenerClass(Object x) {
-        this.advancedOptions.traceListenerClass = x;
-    }
-
-    void traceListenerOutputFile(Object x) {
-        this.advancedOptions.traceListenerOutputFile = x;
-    }
-
-    void traceOptimizerDecisions(Object x) {
-        this.advancedOptions['trace-optimizer-decisions'] = x;
-    }
-
-    void treeModel(Object x) {
-        this.advancedOptions.treeModel = x;
-    }
-
-    void treeModelName(Object x) {
-        this.advancedOptions.treeModelName = x;
-    }
-
-    void unparsedTextURIResolver(Object x) {
-        this.advancedOptions.unparsedTextURIResolver = x;
-    }
-
-    void unparsedTextURIResolverClass(Object x) {
-        this.advancedOptions.unparsedTextURIResolverClass = x;
-    }
-
-    void uriResolverClass(Object x) {
-        this.advancedOptions.uriResolverClass = x;
-    }
-
-    void usePiDisableOutputEscaping(Object x) {
-        this.advancedOptions['use-pi-disable-output-escaping'] = x;
-    }
-
-    void useTypedValueCache(Object x) {
-        this.advancedOptions['use-typed-value-cache'] = x;
-    }
-
-    void useXsiSchemaLocation(Object x) {
-        this.advancedOptions.useXsiSchemaLocation = x;
-    }
-
-    void validationComments(Object x) {
-        this.advancedOptions['validation-comments'] = x;
-    }
-
-    void validationWarnings(Object x) {
-        this.advancedOptions['validation-warnings'] = x;
-    }
-
-    void versionWarning(Object x) {
-        this.advancedOptions['version-warning'] = x;
-    }
-
-    void xincludeAware(Object x) {
-        this.advancedOptions['xinclude-aware'] = x;
-    }
-
-    void xmlVersion(Object x) {
-        this.advancedOptions['xml-version'] = x;
-    }
-
-    void xqueryAllowUpdate(Object x) {
-        this.advancedOptions.xqueryAllowUpdate = x;
-    }
-
-    void xqueryConstructionMode(Object x) {
-        this.advancedOptions.xqueryConstructionMode = x;
-    }
-
-    void xqueryDefaultElementNamespace(Object x) {
-        this.advancedOptions.xqueryDefaultElementNamespace = x;
-    }
-
-    void xqueryDefaultFunctionNamespace(Object x) {
-        this.advancedOptions.xqueryDefaultFunctionNamespace = x;
-    }
-
-    void xqueryEmptyLeast(Object x) {
-        this.advancedOptions.xqueryEmptyLeast = x;
-    }
-
-    void xqueryInheritNamespaces(Object x) {
-        this.advancedOptions.xqueryInheritNamespaces = x;
-    }
-
-    void xqueryMultipleModuleImports(Object x) {
-        this.advancedOptions.xqueryMultipleModuleImports = x;
-    }
-
-    void xqueryPreserveBoundarySpace(Object x) {
-        this.advancedOptions.xqueryPreserveBoundarySpace = x;
-    }
-
-    void xqueryPreserveNamespaces(Object x) {
-        this.advancedOptions.xqueryPreserveNamespaces = x;
-    }
-
-    void xqueryRequiredContextItemType(Object x) {
-        this.advancedOptions.xqueryRequiredContextItemType = x;
-    }
-
-    void xquerySchemaAware(Object x) {
-        this.advancedOptions.xquerySchemaAware = x;
-    }
-
-    void xqueryStaticErrorListenerClass(Object x) {
-        this.advancedOptions.xqueryStaticErrorListenerClass = x;
-    }
-
-    void xqueryVersion(Object x) {
-        this.advancedOptions.xqueryVersion = x;
-    }
-
-    void xsdVersion(Object x) {
-        this.advancedOptions['xsd-version'] = x;
-    }
-
-    void enableAssertions(Object x) {
-        this.advancedOptions.enableAssertions = x;
-    }
-
-    void xsltSchemaAware(Object x) {
-        this.advancedOptions.xsltSchemaAware = x;
-    }
-
-    void stylesheetErrorListener(Object x) {
-        this.advancedOptions.stylesheetErrorListener = x;
-    }
-
-    void stylesheetURIResolver(Object x) {
-        this.advancedOptions.stylesheetURIResolver = x;
-    }
-
-    void xsltVersion(Object x) {
-        this.advancedOptions.xsltVersion = x;
-    }
-
-    // END ADVANCED OPTIONS
 
     void parameters(Map<String, String> parameters) {
         this.stylesheetParams = parameters
@@ -789,13 +257,27 @@ class SaxonXsltTask extends DefaultTask {
     // and output.
     @Internal
     protected List<String> getCommonArguments() {
-        Map<String, String> commonOptions = this.options.findAll { name, value ->
-            !['input', 'output'].contains(name)
-        }.asImmutable()
+        Map<String, String> commonOptions = [:]
+        SaxonPluginConfiguration.instance.configuredOptions().findAll { name, value ->
+            commonOptions[name] = value
+        }
+        this.configuredOptions().findAll { name, value ->
+            commonOptions[name] = value
+        }
+        if (this.options.stylesheet) {
+            commonOptions["stylesheet"] = this.options.stylesheet
+        }
+
+        Map<String, String> advancedOptions = [:]
+        SaxonPluginConfiguration.instance.configuredAdvancedOptions().findAll { name, value ->
+            advancedOptions[name] = value
+        }
+        this.configuredAdvancedOptions().findAll { name, value ->
+            advancedOptions[name] = value
+        }
 
         (commonOptions.inject(this.defaultArguments) { arguments, entry ->
             String defaultArgument = Argument.MAPPING[entry.key]
-
             if (defaultArgument == null) {
                 throw new InvalidUserDataException(
                         "Invalid option: ${entry.key}."
